// client/src/lib/kioskIconData.ts
export const KIOSK_ICON_DATA: KioskIconSet = {
  "version": 1,
  "viewBox": 24,
  "icons": [
    {
      "id": "burger",
      "label": {
        "he": "המבורגרים",
        "en": "Burgers"
      },
// ...
    },
// ...
};